I thought I'd fixed the issue where two CPU cycles would sometimes occur per VME cycle
Things were seemingly working after I increased the clock to 25MHz from
12.5MHz, but it would read ATA sectors from the CF card that were all
the same non-zero value, which wasn't the correct data at that sector.
The problem in the logic analyzer traces was that CPU_AS would go high,
and then low again, before VME_AS would go high.  This was a problem
especially for writing the sector number to the ATA registers, when the
data bus value would change before the CF card had latched it in.

I tried adding an extra state in the vme logic to stop the signals and
then wait for DTACK to be de-asserted.  I managed to remove it later and
only have the AS/DS/transceiver controls de-assert before going to IDLE.
But I also needed changes to the DSACK logic to de-assert while CPU_AS
goes high, and also prevent it from asserting until VME_AS has also gone
high, which prevents an extra cycle.

And then it was still hanging, which turned out to be SystemBoard's
logic not responding on the rare occasion.  I have yet to fix that
issue, and am using a jumper for bus arbitration for the time being.

And just as I was about to commit, I found that it will bus error when
doing an `ls` command, so it's still not right