k30p logic: works as best as it has, but still can't ls /bin
This uses an inverted clock, which should smooth out the clock a bit
more, compensating for only using 2 gates for the clock circuit.  It
also rearranges the dsack logic a bit.  I'm going to try clocking the
dsack logic, and other more drastic changes to get the last issues out.